refactor(world): split server-world.ts into world/ submodules
src/world/world-data.ts   (~204 lines) - mech/room/map data + constants + lookup helpers
src/world/world-scene.ts  (~398 lines) - session helpers + scene/UI packet builders
src/world/world-handlers.ts (~610 lines) - all command/event handlers
src/server-world.ts       (~654 lines) - login, dispatch, connection, factory
