+++ src/server-world.ts
@@ -30,12 +30,10 @@ import {
   buildLoginRequest,
   buildSyncAck,
   buildWelcomePacket,
-  buildCombatWelcomePacket,
 } from './protocol/auth.js';
 import {
   buildCmd36MessageViewPacket,
   buildCmd37OpenComposePacket,
-  buildMenuDialogPacket,
   parseClientCmd4,
   parseClientCmd5SceneAction,
   parseClientCmd10MapReply,
@@ -46,22 +44,14 @@ import {
 } from './protocol/game.js';
 import {
   buildCmd3BroadcastPacket,
-  buildCmd10RoomPresenceSyncPacket,
-  buildCmd11PlayerEventPacket,
-  buildCmd13PlayerArrivalPacket,
-  buildCmd14PersonnelRecordPacket,
-  buildCmd43SolarisMapPacket,
-  buildCmd48KeyedTripleStringListPacket,
   buildCmd4SceneInitPacket,
   buildCmd5CursorNormalPacket,
   buildCmd6CursorBusyPacket,
+  buildCmd10RoomPresenceSyncPacket,
 } from './protocol/world.js';
 import { PlayerRegistry, ClientSession } from './state/players.js';
 import { launchRegistry } from './state/launch.js';
-import { loadMechs } from './data/mechs.js';
-import { loadSolarisRooms, WorldRoom, loadWorldMap, WorldMapRoom } from './data/maps.js';
 import {
-  storeMessage,
   claimUndeliveredMessages,
   markDelivered,
 } from './db/messages.js';
@@ -70,1090 +60,43 @@ import { CaptureLogger } from './util/capture.js';
 import { ARIES_KEEPALIVE_INTERVAL_MS, SOCKET_IDLE_TIMEOUT_MS } from './config.js';
 
 import {
-  buildCmd72LocalBootstrapPacket,
-  buildCmd65PositionSyncPacket,
-  MOTION_NEUTRAL,
-} from './protocol/combat.js';
-
-// ── Shared mech catalog (same on-disk data as lobby) ─────────────────────────
-// Loaded once at module import time.  Provides a fallback when a player's
-// launch record is absent (e.g. direct connection to world port in tests).
-let WORLD_MECHS: ReturnType<typeof loadMechs>;
-try {
-  WORLD_MECHS = loadMechs();
-} catch (err) {
-  const msg = err instanceof Error ? (err.stack ?? err.message) : String(err);
-  // Non-fatal: the world server still starts; FALLBACK_MECH_ID is used instead.
-  process.stderr.write(`[world] WARNING: failed to load mechs: ${msg}\n`);
-  WORLD_MECHS = [];
-}
-
-/** Mech ID used when the player's launch record is missing. */
-const FALLBACK_MECH_ID = WORLD_MECHS.length > 0 ? WORLD_MECHS[0].id : 0;
-
-
-/** Fast lookup from mech ID to MechEntry (for extraCritCount etc.). */
-const WORLD_MECH_BY_ID = new Map(WORLD_MECHS.map(m => [m.id, m]));
-
-// Default arena name shown in the window title.
-const DEFAULT_SCENE_NAME = 'Solaris Arena';
-const WELCOME_TEXT       = 'Welcome to the game world.';
-const DEFAULT_MAP_ROOM_ID = 146; // Solaris Starport
-
-// ── Solaris room model ────────────────────────────────────────────────────────
-// Loaded at startup from SOLARIS.MAP via parseMapFile().  Falls back to a
-// hardcoded list when the proprietary map asset is absent so the server still
-// starts during development without the full game data.
-
-/** Hardcoded fallback used when SOLARIS.MAP is not present. */
-const SOLARIS_FALLBACK_ROOMS: WorldRoom[] = [
-  { roomId: 146, name: 'Solaris Starport',       flags: 0, centreX: 0, centreY: 0, sceneIndex: 0 },
-  { roomId: 147, name: 'Ishiyama Arena',          flags: 0, centreX: 0, centreY: 0, sceneIndex: 1 },
-  { roomId: 148, name: 'Government House',        flags: 0, centreX: 0, centreY: 0, sceneIndex: 2 },
-  { roomId: 149, name: 'White Lotus',             flags: 0, centreX: 0, centreY: 0, sceneIndex: 3 },
-  { roomId: 150, name: 'Waterfront',              flags: 0, centreX: 0, centreY: 0, sceneIndex: 4 },
-  { roomId: 151, name: 'Kobe Slums',              flags: 0, centreX: 0, centreY: 0, sceneIndex: 5 },
-  { roomId: 152, name: 'Steiner Stadium',         flags: 0, centreX: 0, centreY: 0, sceneIndex: 6 },
-  { roomId: 153, name: 'Lyran Building',          flags: 0, centreX: 0, centreY: 0, sceneIndex: 7 },
-  { roomId: 154, name: 'Chahar Park',             flags: 0, centreX: 0, centreY: 0, sceneIndex: 8 },
-  { roomId: 155, name: 'Riverside',               flags: 0, centreX: 0, centreY: 0, sceneIndex: 9 },
-  { roomId: 156, name: 'Black Throne',            flags: 0, centreX: 0, centreY: 0, sceneIndex: 10 },
-  { roomId: 157, name: 'Factory',                 flags: 0, centreX: 0, centreY: 0, sceneIndex: 11 },
-  { roomId: 158, name: 'Marik Tower',             flags: 0, centreX: 0, centreY: 0, sceneIndex: 12 },
-  { roomId: 159, name: 'Allman',                  flags: 0, centreX: 0, centreY: 0, sceneIndex: 13 },
-  { roomId: 160, name: 'Riverfront',              flags: 0, centreX: 0, centreY: 0, sceneIndex: 14 },
-  { roomId: 161, name: 'Wasteland',               flags: 0, centreX: 0, centreY: 0, sceneIndex: 15 },
-  { roomId: 162, name: 'Jungle',                  flags: 0, centreX: 0, centreY: 0, sceneIndex: 16 },
-  { roomId: 163, name: "Chancellor's Quarters",   flags: 0, centreX: 0, centreY: 0, sceneIndex: 17 },
-  { roomId: 164, name: 'Middletown',              flags: 0, centreX: 0, centreY: 0, sceneIndex: 18 },
-  { roomId: 165, name: 'Rivertown',               flags: 0, centreX: 0, centreY: 0, sceneIndex: 19 },
-  { roomId: 166, name: 'Maze',                    flags: 0, centreX: 0, centreY: 0, sceneIndex: 20 },
-  { roomId: 167, name: 'Davion Arena',            flags: 0, centreX: 0, centreY: 0, sceneIndex: 21 },
-  { roomId: 168, name: 'Sortek Building',         flags: 0, centreX: 0, centreY: 0, sceneIndex: 22 },
-  { roomId: 169, name: 'Guzman Park',             flags: 0, centreX: 0, centreY: 0, sceneIndex: 23 },
-  { roomId: 170, name: 'Marina',                  flags: 0, centreX: 0, centreY: 0, sceneIndex: 24 },
-  { roomId: 171, name: 'Viewpoint',               flags: 0, centreX: 0, centreY: 0, sceneIndex: 25 },
-  { roomId: 1,   name: 'International Sector',    flags: 0, centreX: 0, centreY: 0, sceneIndex: 26 },
-  { roomId: 2,   name: 'Kobe Sector',             flags: 0, centreX: 0, centreY: 0, sceneIndex: 27 },
-  { roomId: 3,   name: 'Silesia Sector',          flags: 0, centreX: 0, centreY: 0, sceneIndex: 28 },
-  { roomId: 4,   name: 'Montenegro Sector',       flags: 0, centreX: 0, centreY: 0, sceneIndex: 29 },
-  { roomId: 5,   name: 'Cathay Sector',           flags: 0, centreX: 0, centreY: 0, sceneIndex: 30 },
-  { roomId: 6,   name: 'Black Hills Sector',      flags: 0, centreX: 0, centreY: 0, sceneIndex: 31 },
-];
-
-let solarisRooms: WorldRoom[];
-try {
-  const loaded = loadSolarisRooms();
-  if (loaded) {
-    solarisRooms = loaded;
-    process.stderr.write(`[world] loaded ${loaded.length} rooms from SOLARIS.MAP\n`);
-  } else {
-    solarisRooms = SOLARIS_FALLBACK_ROOMS;
-    process.stderr.write('[world] WARNING: SOLARIS.MAP not found — using hardcoded room list\n');
-  }
-} catch (err) {
-  const msg = err instanceof Error ? (err.stack ?? err.message) : String(err);
-  process.stderr.write(`[world] WARNING: failed to parse SOLARIS.MAP: ${msg}\n`);
-  solarisRooms = SOLARIS_FALLBACK_ROOMS;
-}
-
-const SOLARIS_ROOM_BY_ID = new Map<number, WorldRoom>(
-  solarisRooms.map(room => [room.roomId, room]),
-);
-
-// ── World map (navigation graph from world-map.json) ─────────────────────────
-
-let worldMapByRoomId = new Map<number, WorldMapRoom>();
-try {
-  const worldMap = loadWorldMap();
-  if (worldMap) {
-    worldMapByRoomId = new Map(worldMap.rooms.map(r => [r.roomId, r]));
-    process.stderr.write(`[world] loaded world-map.json (${worldMap.rooms.length} rooms)\n`);
-  } else {
-    process.stderr.write('[world] WARNING: world-map.json not found — using provisional linear topology\n');
-  }
-} catch (err) {
-  const msg = err instanceof Error ? (err.stack ?? err.message) : String(err);
-  process.stderr.write(`[world] WARNING: failed to parse world-map.json: ${msg}\n`);
-}
-
-const ALL_ROSTER_LIST_ID = 0x3F4;
-// 0x3E8 (1000) is reserved by the client for its own local "Personal inquiry on:"
-// submenu (FUN_00412980).  Sending Cmd7 with that listId triggers special client
-// handling that ignores our payload and uses a garbage internal target_id.  Use
-// any non-reserved positive integer instead (see RESEARCH.md §11 avoid-list).
-const INQUIRY_MENU_ID    = 0x3F3;  // 1011 — safe, not in client avoid-list
-const PERSONNEL_LIST_ID  = 0x3F2;
-const PERSONNEL_MORE_ID  = 0x95;
-const SOLARIS_TRAVEL_CONTEXT_ID = 0xC6;
-let nextWorldRosterId    = 1;
-const worldCaptures      = new Map<string, CaptureLogger>();
-
-// ── Helpers ───────────────────────────────────────────────────────────────────
-
-function send(socket: net.Socket, pkt: Buffer, capture: CaptureLogger, label: string): void {
-  capture.logSend(pkt, label);
-  socket.write(pkt);
-}
-
-function sendToWorldSession(session: ClientSession, pkt: Buffer, label: string): void {
-  if (session.socket.destroyed || !session.socket.writable) return;
-  worldCaptures.get(session.id)?.logSend(pkt, label);
-  session.socket.write(pkt);
-}
-
-/**
- * Advance and return the session's outgoing sequence number.
- * Valid range: 0–42 (FUN_0040C2A0: val > 42 → treated as ACK request, not data).
- */
-function nextSeq(session: ClientSession): number {
-  const s = session.serverSeq;
-  session.serverSeq = (session.serverSeq + 1) % 43;
-  return s;
-}
-
-function getDisplayName(session: ClientSession): string {
-  const raw = String((session.displayName ?? session.username) || 'Pilot');
-  const withoutEsc = raw.replace(/[\x00-\x1F\x7F]/g, '');
-  const latin1 = Buffer.from(withoutEsc, 'latin1').subarray(0, 84).toString('latin1');
-  return latin1 || 'Pilot';
-}
-
-function mapRoomKey(roomId: number): string {
-  return `map_room_${roomId}`;
-}
-
-function getSolarisRoomInfo(roomId: number): WorldRoom {
-  return SOLARIS_ROOM_BY_ID.get(roomId) ?? SOLARIS_ROOM_BY_ID.get(DEFAULT_MAP_ROOM_ID)!;
-}
-
-function getSolarisSceneIndex(roomId: number): number {
-  return getSolarisRoomInfo(roomId).sceneIndex;
-}
-
-function getSolarisRoomName(roomId: number): string {
-  return getSolarisRoomInfo(roomId).name;
-}
-
-function uniqueRoomIds(roomIds: number[]): number[] {
-  return [...new Set(roomIds)].filter(roomId => SOLARIS_ROOM_BY_ID.has(roomId));
-}
-
-/**
- * Return up to 4 exit room IDs for a given room.
- *
- * SOLARIS.MAP stores pixel positions on the visual travel-map bitmap and
- * room descriptions.  It does NOT encode room-to-room connections.
- * IS.MAP similarly stores star-system positions on the IS overview bitmap.
- * Neither file contains a navigation graph.
- *
- * The original server's room connection table has not yet been RE'd.
- * Until it is, use a provisional linear topology: room 146 is the Solaris
- * hub, each room connects back to the hub and to its immediate neighbours
- * in the loaded room list, with sector-row links for Solaris district rooms.
- */
-function getSolarisRoomExits(roomId: number): number[] {
-  // Use world-map.json if loaded.
-  const mapRoom = worldMapByRoomId.get(roomId);
-  if (mapRoom) {
-    const { north, south, east, west } = mapRoom.exits;
-    // Slot order: 0=N 1=S 2=E 3=W.  Skip nulls (no exit in that direction).
-    return [north, south, east, west].filter((id): id is number => id !== null);
-  }
-
-  // Provisional linear topology fallback.
-  if (roomId === 146) return [147, 152, 157, 162];
-
-  const index = solarisRooms.findIndex(r => r.roomId === roomId);
-  const exits = [146];
-
-  if (index > 0) exits.push(solarisRooms[index - 1].roomId);
-  if (index >= 0 && index < solarisRooms.length - 1) exits.push(solarisRooms[index + 1].roomId);
-
-  if (roomId >= 147 && roomId <= 171) {
-    const sectorOffset = Math.floor((roomId - 147) / 5);
-    exits.push(solarisRooms[26 + Math.min(sectorOffset, 5)].roomId);
-  }
-
-  return uniqueRoomIds(exits).filter(exit => exit !== roomId).slice(0, 4);
-}
-
-/**
- * Return the Cmd4 location icon ID (mechId) for a room.
- * Uses the icon field from world-map.json if present; falls back to sceneIndex.
- */
-function getSolarisRoomIcon(roomId: number): number {
-  const mapRoom = worldMapByRoomId.get(roomId);
-  if (mapRoom?.icon !== null && mapRoom?.icon !== undefined) return mapRoom.icon;
-  return getSolarisSceneIndex(roomId);
-}
-
-function getPresenceStatus(session: ClientSession): number {
-  return session.worldPresenceStatus ?? 5;
-}
-
-function getComstarId(session: ClientSession): number {
-  if (session.accountId !== undefined) {
-    return 100000 + session.accountId;
-  }
-  return 900000 + (session.worldRosterId ?? 0);
-}
-
-function getPresenceLocation(session: ClientSession): string {
-  const roomId = session.worldMapRoomId;
-  const status = getPresenceStatus(session);
-  const room = roomId === undefined ? 'world' : getSolarisRoomName(roomId);
-  if (status <= 5) return `Standing in ${room}`;
-  if (status <= 12) return `Booth ${status - 5} in ${room}`;
-  return `Status ${status}`;
-}
-
-function currentRoomPresenceEntries(players: PlayerRegistry, session: ClientSession) {
-  if (session.worldRosterId === undefined) {
-    return [];
-  }
-
-  const entries = [
-    {
-      rosterId: session.worldRosterId,
-      status:   getPresenceStatus(session),
-      callsign: getDisplayName(session),
-    },
-  ];
-
-  for (const other of players.inRoom(session.roomId)) {
-    if (
-      other.id === session.id ||
-      other.phase !== 'world' ||
-      !other.worldInitialized ||
-      other.worldRosterId === undefined ||
-      other.socket.destroyed
-    ) {
-      continue;
-    }
-
-    entries.push({
-      rosterId: other.worldRosterId,
-      status:   getPresenceStatus(other),
-      callsign: getDisplayName(other),
-    });
-  }
-
-  return entries;
-}
-
-function findWorldTargetBySelectionId(
-  players: PlayerRegistry,
-  targetId: number,
-): ClientSession | undefined {
-  return players.worldSessions().find(other =>
-    getComstarId(other) === targetId || other.worldRosterId === targetId,
-  );
-}
-
-function buildAllRosterEntries(players: PlayerRegistry) {
-  return players.worldSessions()
-    .slice()
-    .sort((a, b) => getComstarId(a) - getComstarId(b))
-    .map(other => ({
-      itemId: getComstarId(other),
-      col1:   getDisplayName(other),
-      col2:   getSolarisRoomName(other.worldMapRoomId ?? DEFAULT_MAP_ROOM_ID),
-      col3:   getPresenceLocation(other),
-    }));
-}
-
-function buildPersonnelRecordLines(target: ClientSession, page: number): string[] {
-  if (page <= 1) {
-    return [
-      // The client's Cmd14 header always shows the querying user's own callsign
-      // as "Handle" (it reads from the room-roster selection cursor, which
-      // defaults to self).  We have no wire field that overrides it, so we
-      // repeat the correct handle as the first body line.
-      `Handle   : ${getDisplayName(target)}`,
-      `House    : ${target.allegiance ?? 'Unaffiliated'}`,
-      `Sector   : ${getSolarisRoomName(target.worldMapRoomId ?? DEFAULT_MAP_ROOM_ID)}`,
-      `Location : ${getPresenceLocation(target)}`,
-      'Status   : Online',
-      `ComStar  : ${getComstarId(target)}`,
-    ];
-  }
-
-  return [
-    'Stable   : Independent',
-    `Mech ID  : ${target.selectedMechId ?? FALLBACK_MECH_ID}`,
-    `Roster   : ${target.worldRosterId ?? 0}`,
-    'Standing : 0',
-    'Winnings : 0 cb',
-    'Record   : Prototype page 2',
-  ];
-}
-
-function buildComstarDeliveryText(senderName: string, text: string): string {
-  const raw = `ComStar message from ${senderName}\\${text}`;
-  let trimmed = raw.replace(/\x1b/g, '?');
-  while (Buffer.byteLength(trimmed, 'latin1') > (85 * 85 - 1)) {
-    trimmed = trimmed.slice(0, -1);
-  }
-  return trimmed;
-}
-
-function sendSolarisTravelMap(
-  session: ClientSession,
-  connLog: Logger,
-  capture: CaptureLogger,
-): void {
-  const currentRoomId = session.worldMapRoomId ?? DEFAULT_MAP_ROOM_ID;
-  connLog.info('[world] sending Cmd43 Solaris travel map: currentRoomId=%d', currentRoomId);
-  send(
-    session.socket,
-    buildCmd43SolarisMapPacket(
-      {
-        contextId: SOLARIS_TRAVEL_CONTEXT_ID,
-        currentRoomId,
-      },
-      nextSeq(session),
-    ),
-    capture,
-    'CMD43_SOLARIS_MAP',
-  );
-}
-
-function buildSceneInitForSession(session: ClientSession) {
-  const roomId = session.worldMapRoomId ?? DEFAULT_MAP_ROOM_ID;
-  const sceneIndex = getSolarisSceneIndex(roomId);
-
-  // Build a 4-slot array (N=0, S=1, E=2, W=3) preserving direction positions.
-  // When using world-map.json the exits array already has nulls for empty slots.
-  // When using the fallback, compact exits are back-filled starting at slot 0.
-  const mapRoom = worldMapByRoomId.get(roomId);
-  let slottedExits: (number | null)[];
-  if (mapRoom) {
-    const { north, south, east, west } = mapRoom.exits;
-    slottedExits = [north, south, east, west];
-  } else {
-    const exits = getSolarisRoomExits(roomId);
-    slottedExits = [
-      exits[0] ?? null,
-      exits[1] ?? null,
-      exits[2] ?? null,
-      exits[3] ?? null,
-    ];
-  }
-
-  const exitMask = slottedExits.reduce<number>(
-    (mask, id, slot) => (id !== null ? mask | (1 << slot) : mask),
-    0,
-  );
-
-  return buildCmd4SceneInitPacket(
-    {
-      sessionFlags:     0x30 | exitMask,
-      playerScoreSlot:  sceneIndex,
-      playerMechId:     getSolarisRoomIcon(roomId),
-      opponents:        slottedExits.map(exitRoomId => {
-        if (exitRoomId === null) return null as unknown as { type: number; mechId: number };
-        return { type: getSolarisSceneIndex(exitRoomId), mechId: getSolarisRoomIcon(exitRoomId) };
-      }),
-      callsign:         getDisplayName(session),
-      sceneName:        getSolarisRoomName(roomId),
-      // The client special-cases the first Cmd4 option button (0x100) as local Help.
-      // Put server actions at 0x101+ so FUN_00413790 emits cmd 5 with the type byte.
-      arenaOptions:     [
-        { type: 0, label: 'Help' },
-        { type: 4, label: 'Travel' },
-      ],
-    },
-    nextSeq(session),
-  );
-}
-
-function sendSceneRefresh(
-  players: PlayerRegistry,
-  session: ClientSession,
-  connLog: Logger,
-  capture: CaptureLogger,
-  message: string,
-): void {
-  send(session.socket, buildCmd6CursorBusyPacket(nextSeq(session)), capture, 'CMD6_BUSY');
-  send(session.socket, buildSceneInitForSession(session), capture, 'CMD4_SCENE_REFRESH');
-
-  const roomPresenceEntries = currentRoomPresenceEntries(players, session);
-  connLog.info('[world] sending Cmd10 RoomPresenceSync (%d entries)', roomPresenceEntries.length);
-  send(
-    session.socket,
-    buildCmd10RoomPresenceSyncPacket(roomPresenceEntries, nextSeq(session)),
-    capture,
-    'CMD10_ROOM_SYNC',
-  );
-  send(
-    session.socket,
-    buildCmd3BroadcastPacket(message, nextSeq(session)),
-    capture,
-    'CMD3_TRAVEL_COMPLETE',
-  );
-  send(session.socket, buildCmd5CursorNormalPacket(nextSeq(session)), capture, 'CMD5_NORMAL');
-}
-
-function sendAllRosterList(
-  players: PlayerRegistry,
-  session: ClientSession,
-  connLog: Logger,
-  capture: CaptureLogger,
-): void {
-  const entries = buildAllRosterEntries(players);
-  connLog.info('[world] sending Cmd48 all-roster list (%d entries)', entries.length);
-  send(
-    session.socket,
-    buildCmd48KeyedTripleStringListPacket(
-      ALL_ROSTER_LIST_ID,
-      'All Personnel Online',
-      entries,
-      nextSeq(session),
-    ),
-    capture,
-    'CMD48_ALL_ROSTER',
-  );
-}
-
-function sendInquiryMenu(
-  session: ClientSession,
-  target: ClientSession,
-  connLog: Logger,
-  capture: CaptureLogger,
-): void {
-  const targetId = getComstarId(target);
-  session.worldInquiryTargetId = targetId;
-  session.worldInquiryPage = undefined;
-
-  connLog.info(
-    '[world] sending inquiry submenu: target=%d handle="%s"',
-    targetId,
-    getDisplayName(target),
-  );
-  send(
-    session.socket,
-    buildMenuDialogPacket(
-      INQUIRY_MENU_ID,
-      'Personal inquiry on:',
-      ['Send a ComStar message', 'Access personnel data'],
-      nextSeq(session),
-    ),
-    capture,
-    'CMD7_INQUIRY_MENU',
-  );
-}
-
-function sendPersonnelRecord(
-  players: PlayerRegistry,
-  session: ClientSession,
-  targetId: number,
-  page: number,
-  connLog: Logger,
-  capture: CaptureLogger,
-): void {
-  const target = findWorldTargetBySelectionId(players, targetId);
-  if (!target) {
-    connLog.warn('[world] personnel record target not found: id=%d page=%d', targetId, page);
-    send(
-      session.socket,
-      buildCmd14PersonnelRecordPacket(
-        {
-          comstarId: targetId,
-          battlesToDate: 0,
-          lines: ['Status   : Offline', 'Record   : Unavailable', '', '', '', ''],
-        },
-        nextSeq(session),
-      ),
-      capture,
-      'CMD14_PERSONNEL_OFFLINE',
-    );
-    return;
-  }
-
-  const resolvedTargetId = getComstarId(target);
-  session.worldInquiryTargetId = resolvedTargetId;
-  session.worldInquiryPage = page;
-
-  // The client's Cmd14 handler looks up the handle for the record in the room
-  // presence table (seeded by Cmd10/Cmd13), which is keyed by worldRosterId.
-  // Sending getComstarId (100000+accountId) as comstarId results in a lookup
-  // miss → "Handle = null" and the client falls back to its own callsign.
-  // The real ComStar ID is already shown in the body lines ('ComStar  : N').
-  const presenceId = target.worldRosterId ?? 0;
-
-  connLog.info(
-    '[world] sending Cmd14 personnel record: presenceId=%d handle="%s" page=%d',
-    presenceId,
-    getDisplayName(target),
-    page,
-  );
-  send(
-    session.socket,
-    buildCmd14PersonnelRecordPacket(
-      {
-        comstarId:     presenceId,
-        battlesToDate: 0,
-        lines:         buildPersonnelRecordLines(target, page),
-      },
-      nextSeq(session),
-    ),
-    capture,
-    page <= 1 ? 'CMD14_PERSONNEL_P1' : 'CMD14_PERSONNEL_P2',
-  );
-}
-
-function handleComstarTextReply(
-  players: PlayerRegistry,
-  session: ClientSession,
-  dialogId: number,
-  text: string,
-  connLog: Logger,
-  capture: CaptureLogger,
-): void {
-  const clean = text.replace(/\x1b/g, '?').replace(/\s+/g, ' ').trim();
-  if (clean.length === 0) {
-    connLog.warn('[world] cmd-21 ComStar text ignored (empty)');
-    send(
-      session.socket,
-      buildCmd3BroadcastPacket('ComStar message not sent: empty text.', nextSeq(session)),
-      capture,
-      'CMD3_COMSTAR_EMPTY',
-    );
-    return;
-  }
-
-  const senderName   = getDisplayName(session);
-  const senderComstarId = getComstarId(session);
-  const formattedBody   = buildComstarDeliveryText(senderName, clean);
-
-  const target = findWorldTargetBySelectionId(players, dialogId);
-  if (target) {
-    // Recipient is online — deliver immediately.
-    const targetName = getDisplayName(target);
-    connLog.info(
-      '[world] cmd-21 ComStar (online): from="%s" to="%s" target=%d text=%j',
-      senderName, targetName, dialogId, clean,
-    );
-    sendToWorldSession(
-      target,
-      buildCmd36MessageViewPacket(senderComstarId, formattedBody, nextSeq(target)),
-      'CMD36_COMSTAR_DELIVERY',
-    );
-    send(
-      session.socket,
-      buildCmd3BroadcastPacket(`ComStar sent to ${targetName}.`, nextSeq(session)),
-      capture,
-      'CMD3_COMSTAR_ACK',
-    );
-    return;
-  }
-
-  // Recipient is offline (or their session ended between roster fetch and now).
-  // comstarId = 100_000 + accountId for authenticated players;
-  // 900_000 + worldRosterId for anonymous sessions (cannot persist).
-  const recipientAccountId =
-    dialogId > 100_000 && dialogId < 900_000 ? dialogId - 100_000 : undefined;
-  const senderAccountId = session.accountId;
-
-  if (senderAccountId !== undefined && recipientAccountId !== undefined) {
-    connLog.info(
-      '[world] cmd-21 ComStar (offline): from=%d to account=%d text=%j — persisting',
-      senderAccountId, recipientAccountId, clean,
-    );
-    storeMessage(senderAccountId, recipientAccountId, senderComstarId, formattedBody)
-      .then(() => {
-        connLog.info('[world] ComStar message stored for offline delivery (account=%d)', recipientAccountId);
-        if (!session.socket.destroyed && session.socket.writable) {
-          send(
-            session.socket,
-            buildCmd3BroadcastPacket('ComStar message queued for offline delivery.', nextSeq(session)),
-            capture,
-            'CMD3_COMSTAR_QUEUED',
-          );
-        }
-      })
-      .catch((err: unknown) => {
-        const msg = err instanceof Error ? err.message : String(err);
-        connLog.error('[world] failed to store offline ComStar: %s', msg);
-        if (!session.socket.destroyed && session.socket.writable) {
-          send(
-            session.socket,
-            buildCmd3BroadcastPacket('ComStar delivery failed \u2014 please try again.', nextSeq(session)),
-            capture,
-            'CMD3_COMSTAR_FAIL',
-          );
-        }
-      });
-  } else {
-    connLog.warn(
-      '[world] cmd-21 ComStar target unavailable and cannot persist: dialogId=%d senderAccId=%s',
-      dialogId, senderAccountId,
-    );
-    send(
-      session.socket,
-      buildCmd3BroadcastPacket('ComStar target unavailable.', nextSeq(session)),
-      capture,
-      'CMD3_COMSTAR_MISSING',
-    );
-  }
-}
-
-function nextAvailableBooth(players: PlayerRegistry, roomId: string, excludeId: string): number {
-  const occupied = new Set<number>();
-  for (const other of players.inRoom(roomId)) {
-    if (
-      other.id === excludeId ||
-      other.phase !== 'world' ||
-      !other.worldInitialized ||
-      other.socket.destroyed
-    ) {
-      continue;
-    }
-
-    const booth = getPresenceStatus(other) - 5;
-    if (booth > 0) occupied.add(booth);
-  }
-
-  for (let booth = 1; booth <= 7; booth += 1) {
-    if (!occupied.has(booth)) return booth;
-  }
-
-  return 1;
-}
-
-function updateRoomPresenceStatus(
-  players: PlayerRegistry,
-  session: ClientSession,
-  status: number,
-  connLog: Logger,
-): void {
-  if (
-    !session.roomId ||
-    session.worldRosterId === undefined ||
-    !session.worldInitialized
-  ) {
-    return;
-  }
-
-  if (getPresenceStatus(session) === status) {
-    connLog.debug('[world] room presence unchanged: rosterId=%d status=%d', session.worldRosterId, status);
-    return;
-  }
-
-  session.worldPresenceStatus = status;
-  const callsign = getDisplayName(session);
-  for (const other of players.inRoom(session.roomId)) {
-    if (
-      other.phase !== 'world' ||
-      !other.worldInitialized ||
-      other.socket.destroyed
-    ) {
-      continue;
-    }
-    sendToWorldSession(
-      other,
-      buildCmd11PlayerEventPacket(session.worldRosterId, status, callsign, nextSeq(other)),
-      'CMD11_STATUS_UPDATE',
-    );
-  }
-
-  connLog.info(
-    '[world] room presence update: rosterId=%d status=%d callsign="%s"',
-    session.worldRosterId,
-    status,
-    callsign,
-  );
-}
-
-function handleRoomMenuSelection(
-  players: PlayerRegistry,
-  session: ClientSession,
-  selection: number,
-  connLog: Logger,
-  capture: CaptureLogger,
-): void {
-  if (selection === 1) {
-    connLog.info('[world] room menu: all-roster request');
-    sendAllRosterList(players, session, connLog, capture);
-    return;
-  }
-
-  if (selection === 0) {
-    const booth = nextAvailableBooth(players, session.roomId, session.id);
-    connLog.info('[world] room menu: new booth requested -> booth %d', booth);
-    updateRoomPresenceStatus(players, session, 5 + booth, connLog);
-    return;
-  }
-
-  if (selection === 2) {
-    connLog.info('[world] room menu: stand requested');
-    updateRoomPresenceStatus(players, session, 5, connLog);
-    return;
-  }
-
-  const booth = selection - 2;
-  if (booth < 1 || booth > 7) {
-    connLog.warn('[world] room menu: unsupported booth selection=%d', selection);
-    return;
-  }
-
-  connLog.info('[world] room menu: join booth %d', booth);
-  updateRoomPresenceStatus(players, session, 5 + booth, connLog);
-}
-
-/**
- * Send the combat entry bootstrap sequence after the player types "/fight".
- *
- * Protocol order (CONFIRMED by Ghidra RE of Main_ModePacketDispatch_v123):
- *   1. MMC SYNC — raw ARIES packet; triggers client RPS→combat dispatch-table
- *      switch.  Client calls Main_SetModeName_v123(1) + Combat_InitMode_v123()
- *      (loads scenes.dat locally — no server data required for that step).
- *   2. Cmd72   — local-bootstrap game frame using combat CRC seed (0x0A5C45).
- *      Seeds scenario title, terrain, identity strings, spawn coords, and the
- *      local mech damage state.  remainingActorCount=0 → solo arena (no bots).
- *
- * Unresolved assumptions (safe defaults used):
- *   • terrainId / terrainResourceId — 1/0 chosen; live capture needed.
- *   • identity2..4 — empty; purpose in client UI unconfirmed.
- *   • headingBias  — 0 (MOTION_NEUTRAL added by encoder); live capture needed.
- *   • globalA/B/C  — 0; purpose unlabelled in Ghidra.
- */
-function sendCombatBootstrapSequence(
-  session: ClientSession,
-  connLog: Logger,
-  capture: CaptureLogger,
-): void {
-  const { socket } = session;
-  const mechId   = session.selectedMechId ?? FALLBACK_MECH_ID;
-  const callsign = getDisplayName(session);
-
-  // Look up the mech's extra crit count (confirmed by RE of
-  // Combat_ReadLocalActorMechState_v123 @ 0x004456c0 — the client reads
-  // extraCritCount + 21 bytes from the packet, where extraCritCount comes from
-  // the mech's .MEC file at offset 0x3c after decryption).
-  const mechEntry       = WORLD_MECH_BY_ID.get(mechId);
-  const extraCritCount  = mechEntry?.extraCritCount ?? 0;
-  const critBytes       = Math.max(0, extraCritCount + 21);
-
-  // 1. MMC SYNC — plain ARIES packet; no game-frame CRC.
-  send(socket, buildCombatWelcomePacket(), capture, 'COMBAT_WELCOME_MMC');
-
-  // Switch phase *before* sending combat game frames so that any inbound
-  // frames that arrive immediately use the correct CRC seed.
-  session.phase = 'combat';
-
-  // 2. Cmd72 — local bootstrap (combat CRC seed applied by buildGamePacket).
-  const cmd72 = buildCmd72LocalBootstrapPacket(
-    {
-      scenarioTitle:      DEFAULT_SCENE_NAME,
-      localSlot:          0,
-      unknownByte0:       0,
-      terrainId:          1,      // ASSUMPTION: default terrain set
-      terrainResourceId:  0,      // ASSUMPTION: no additional resource
-      terrainPoints:      [],
-      arenaPoints:        [],
-      globalA:            0,
-      globalB:            0,
-      globalC:            0,
-      headingBias:        0,      // ASSUMPTION: 0 → MOTION_NEUTRAL after encode
-      identity0:          callsign.substring(0, 11),
-      identity1:          callsign.substring(0, 31),
-      identity2:          '',     // ASSUMPTION: mech type or empty
-      identity3:          '',     // ASSUMPTION: house or empty
-      identity4:          '',     // ASSUMPTION: unknown; empty safe
-      statusByte:         0,
-      initialX:           0,
-      initialY:           0,
-      extraType2Values:   [],
-      remainingActorCount: 0,     // solo arena — no remote actors
-      unknownType1Raw:    MOTION_NEUTRAL,
-      mech: {
-        mechId,
-        critStateExtraCount:  extraCritCount,
-        criticalStateBytes:   Array<number>(critBytes).fill(0),
-        extraStateBytes:      [],
-        armorLikeStateBytes:  Array<number>(11).fill(0),  // full armor
-        internalStateBytes:   Array<number>(8).fill(0),   // full internals
-        ammoStateValues:      [],
-        actorDisplayName:     callsign.substring(0, 31),
-      },
-    },
-    nextSeq(session),
-  );
-
-  connLog.info('[world] sending Cmd72 combat bootstrap (mech_id=%d callsign="%s")', mechId, callsign);
-  send(socket, cmd72, capture, 'CMD72_COMBAT_BOOTSTRAP');
-
-  // 3. Cmd65 — initial position for the local actor at the origin.
-  //    Gives the client something to render immediately after bootstrap.
-  //    facing/throttle/legVel/speedMag = 0 (stationary, no heading).
-  const cmd65 = buildCmd65PositionSyncPacket(
-    { slot: 0, x: 0, y: 0, z: 0, facing: 0, throttle: 0, legVel: 0, speedMag: 0 },
-    nextSeq(session),
-  );
-  send(socket, cmd65, capture, 'CMD65_INITIAL_POSITION');
-
-  session.combatInitialized = true;
-  connLog.info('[world] combat entry complete for "%s"', callsign);
-}
-
-function handleWorldTextCommand(
-  players: PlayerRegistry,
-  session: ClientSession,
-  text: string,
-  connLog: Logger,
-  capture: CaptureLogger,
-): void {
-  const clean = text.replace(/\x1b/g, '?').trim();
-  if (clean.length === 0) {
-    connLog.debug('[world] cmd-4 text ignored (empty)');
-    return;
-  }
-
-  if (clean.toLowerCase() === '/map' || clean.toLowerCase() === '/travel') {
-    sendSolarisTravelMap(session, connLog, capture);
-    return;
-  }
-
-  // /icons [start] — send a fake scene with 4 exit slots showing icons N, N+1,
-  // N+2, N+3.  Used to empirically map icon IDs to their displayed graphics.
-  const iconsMatch = clean.match(/^\/icons(?:\s+(\d+))?$/i);
-  if (iconsMatch) {
-    const base = parseInt(iconsMatch[1] ?? '0', 10);
-    connLog.info('[world] /icons test: base=%d', base);
-    send(
-      session.socket,
-      buildCmd4SceneInitPacket(
-        {
-          sessionFlags:    0x30 | 0x0F,  // all 4 slots enabled
-          playerScoreSlot: 0,
-          playerMechId:    base,
-          opponents: [
-            { type: 0, mechId: base },
-            { type: 0, mechId: base + 1 },
-            { type: 0, mechId: base + 2 },
-            { type: 0, mechId: base + 3 },
-          ],
-          callsign:  getDisplayName(session),
-          sceneName: `Icons ${base}–${base + 3}`,
-          arenaOptions: [
-            { type: 0, label: 'Help' },
-            { type: 4, label: 'Travel' },
-          ],
-        },
-        nextSeq(session),
-      ),
-      capture,
-      'CMD4_ICONS_TEST',
-    );
-    send(
-      session.socket,
-      buildCmd3BroadcastPacket(
-        `N=${base} S=${base+1} E=${base+2} W=${base+3}  (center=${base})  Type /icons ${base+4} for next batch.`,
-        nextSeq(session),
-      ),
-      capture,
-      'CMD3_ICONS_LABEL',
-    );
-    return;
-  }
-  const line = `${getDisplayName(session)}: ${clean}`;
-  connLog.info('[world] cmd-4 text: %s', line);
-
-  const senderStatus = getPresenceStatus(session);
-  const senderInBooth = senderStatus > 5;
-
-  for (const other of players.inRoom(session.roomId)) {
-    if (
-      other.phase !== 'world' ||
-      !other.worldInitialized ||
-      other.socket.destroyed
-    ) {
-      continue;
-    }
-
-    // Booth privacy: booth chat is only heard by occupants of the same booth;
-    // standing chat is only heard by other standing players.
-    const otherStatus = getPresenceStatus(other);
-    if (senderInBooth ? otherStatus !== senderStatus : otherStatus > 5) {
-      continue;
-    }
-
-    sendToWorldSession(other, buildCmd3BroadcastPacket(line, nextSeq(other)), 'CMD3_CHAT_FANOUT');
-  }
-}
-
-function handleMapTravelReply(
-  players: PlayerRegistry,
-  session: ClientSession,
-  contextId: number,
-  selection: number,
-  selectedRoomId: number | undefined,
-  connLog: Logger,
-  capture: CaptureLogger,
-): void {
-  if (selection === 0) {
-    connLog.info('[world] cmd-10 map reply: context=%d cancel', contextId);
-    return;
-  }
-
-  if (selectedRoomId === undefined) {
-    connLog.warn('[world] cmd-10 map reply missing selected room: context=%d selection=%d', contextId, selection);
-    return;
-  }
-
-  if (contextId !== SOLARIS_TRAVEL_CONTEXT_ID) {
-    connLog.warn('[world] cmd-10 map reply unexpected context=%d (expected %d)', contextId, SOLARIS_TRAVEL_CONTEXT_ID);
-    return;
-  }
-
-  if (!SOLARIS_ROOM_BY_ID.has(selectedRoomId)) {
-    connLog.warn('[world] cmd-10 map reply unknown selectedRoomId=%d', selectedRoomId);
-    return;
-  }
-
-  const oldRoomId = session.roomId;
-  const newRoomId = mapRoomKey(selectedRoomId);
-  connLog.info(
-    '[world] cmd-10 map reply: context=%d selection=%d selectedRoomId=%d',
-    contextId,
-    selection,
-    selectedRoomId,
-  );
-
-  if (oldRoomId === newRoomId) {
-    send(
-      session.socket,
-      buildCmd3BroadcastPacket(`Already at room ${selectedRoomId}.`, nextSeq(session)),
-      capture,
-      'CMD3_TRAVEL_ALREADY_THERE',
-    );
-    return;
-  }
-
-  notifyRoomDeparture(players, session, connLog);
-  session.roomId = newRoomId;
-  session.worldMapRoomId = selectedRoomId;
-  session.worldPresenceStatus = 5;
-
-  sendSceneRefresh(
-    players,
-    session,
-    connLog,
-    capture,
-    `Travel complete: ${getSolarisRoomName(selectedRoomId)}.`,
-  );
-  notifyRoomArrival(players, session, connLog);
-}
-
-function handleLocationAction(
-  players: PlayerRegistry,
-  session: ClientSession,
-  slot: number,
-  targetCached: boolean,
-  connLog: Logger,
-  capture: CaptureLogger,
-): void {
-  const currentRoomId = session.worldMapRoomId ?? DEFAULT_MAP_ROOM_ID;
-
-  // Resolve exit by compass slot (0=N 1=S 2=E 3=W).  Must use slotted exits
-  // (nulls preserved) so slot indices match the buttons sent to the client.
-  // getSolarisRoomExits() returns a compact filtered array — do NOT use it here.
-  const mapRoom = worldMapByRoomId.get(currentRoomId);
-  let targetRoomId: number | undefined;
-  if (mapRoom) {
-    const slotted: (number | null)[] = [
-      mapRoom.exits.north, mapRoom.exits.south, mapRoom.exits.east, mapRoom.exits.west,
-    ];
-    targetRoomId = slotted[slot] ?? undefined;
-  } else {
-    // Fallback linear topology densely fills all slots, so compact is fine.
-    targetRoomId = getSolarisRoomExits(currentRoomId)[slot];
-  }
-
-  if (targetRoomId === undefined) {
-    connLog.warn('[world] cmd-23 location action has no exit: room=%d slot=%d cached=%s', currentRoomId, slot, targetCached);
-    send(
-      session.socket,
-      buildCmd3BroadcastPacket('There is no exit in that direction.', nextSeq(session)),
-      capture,
-      'CMD3_LOCATION_NO_EXIT',
-    );
-    return;
-  }
-
-  connLog.info(
-    '[world] cmd-23 location action: room=%d slot=%d cached=%s -> room=%d',
-    currentRoomId,
-    slot,
-    targetCached,
-    targetRoomId,
-  );
-
-  notifyRoomDeparture(players, session, connLog);
-  session.roomId = mapRoomKey(targetRoomId);
-  session.worldMapRoomId = targetRoomId;
-  session.worldPresenceStatus = 5;
-  sendSceneRefresh(
-    players,
-    session,
-    connLog,
-    capture,
-    `Arrived at ${getSolarisRoomName(targetRoomId)}.`,
-  );
-  notifyRoomArrival(players, session, connLog);
-}
-
-function notifyRoomArrival(
-  players: PlayerRegistry,
-  session: ClientSession,
-  connLog: Logger,
-): void {
-  if (!session.roomId || session.worldRosterId === undefined) return;
-  const callsign = getDisplayName(session);
-  for (const other of players.inRoom(session.roomId)) {
-    if (
-      other.id === session.id ||
-      other.phase !== 'world' ||
-      !other.worldInitialized ||
-      other.socket.destroyed
-    ) {
-      continue;
-    }
-    sendToWorldSession(
-      other,
-      buildCmd13PlayerArrivalPacket(session.worldRosterId, callsign, nextSeq(other)),
-      'CMD13_ARRIVAL',
-    );
-  }
-  connLog.info('[world] notified room of arrival: rosterId=%d callsign="%s"', session.worldRosterId, callsign);
-}
-
-function notifyRoomDeparture(
-  players: PlayerRegistry,
-  session: ClientSession,
-  connLog: Logger,
-): void {
-  if (!session.roomId || session.worldRosterId === undefined) return;
-  const callsign = getDisplayName(session);
-  for (const other of players.inRoom(session.roomId)) {
-    if (
-      other.id === session.id ||
-      other.phase !== 'world' ||
-      !other.worldInitialized ||
-      other.socket.destroyed
-    ) {
-      continue;
-    }
-    sendToWorldSession(
-      other,
-      buildCmd11PlayerEventPacket(session.worldRosterId, 0, callsign, nextSeq(other)),
-      'CMD11_DEPARTURE',
-    );
-  }
-  connLog.info('[world] notified room of departure: rosterId=%d callsign="%s"', session.worldRosterId, callsign);
-}
+  worldCaptures,
+  allocateWorldRosterId,
+  DEFAULT_MAP_ROOM_ID,
+  ALL_ROSTER_LIST_ID,
+  INQUIRY_MENU_ID,
+  PERSONNEL_LIST_ID,
+  PERSONNEL_MORE_ID,
+  SOLARIS_TRAVEL_CONTEXT_ID,
+  getSolarisRoomName,
+} from './world/world-data.js';
+import {
+  send,
+  nextSeq,
+  getDisplayName,
+  mapRoomKey,
+  getComstarId,
+  findWorldTargetBySelectionId,
+  sendInquiryMenu,
+  sendPersonnelRecord,
+  buildSceneInitForSession,
+  sendSceneRefresh,
+  sendAllRosterList,
+  sendSolarisTravelMap,
+  currentRoomPresenceEntries,
+} from './world/world-scene.js';
+import {
+  handleComstarTextReply,
+  handleRoomMenuSelection,
+  handleMapTravelReply,
+  handleLocationAction,
+  handleWorldTextCommand,
+  sendCombatBootstrapSequence,
+  notifyRoomArrival,
+  notifyRoomDeparture,
+} from './world/world-handlers.js';
+
+const WELCOME_TEXT = 'Welcome to the game world.';
 
 // ── Login handler ─────────────────────────────────────────────────────────────
 
@@ -1575,7 +518,7 @@ function handleWorldConnection(socket: net.Socket, players: PlayerRegistry, log:
     awaitingMechConfirm: false,
     serverSeq:         0,
     worldInitialized:  false,
-    worldRosterId:     nextWorldRosterId++,
+    worldRosterId:     allocateWorldRosterId(),
     worldPresenceStatus: 5,
   };
   players.add(session);
@@ -1708,3 +651,4 @@ export function startWorldServer(log: Logger, players: PlayerRegistry): net.Serv
 
   return worldServer;
 }
+
